Separate API controller to improve routing performance
